加强异步资源清理并优化表达式计算 - 在 `DatabaseExecutor.Async.cs` 中，为所有异步调用添加了 `ConfigureAwait(false)` 以避免死锁并提高性能。 - 在 `LookupDb.cs` 中引入了一个新方法 `AppendJsonCast` 用于高效的 SQL 参数转换，用高性能的字符串操作方法取代了正则表达式。 - 对 `RepositoryRouter.cs` 进行了重构，使用字典来映射列名到属性名，从而提高查找效率。 - 在 `AggregateTests.cs` 中更新了聚合测试，以确保对求和操作的类型处理正确。 - 在 `LookupDbAppendJsonCastTests.cs` 中为 `AppendJsonCast` 功能添加了单元测试，以验证各种 SQL 参数格式和边缘情况下的行为。

+++ tests/Inkslab.Linq.Tests/AggregateTests.cs
@@ -844,13 +844,13 @@ namespace Inkslab.Linq.Tests
 
             var expectedGroups = allUsers
                 .GroupBy(x => x.IsAdministrator)
-                .Select(g => new { Key = g.Key, SumId = g.Sum(x => x.Id) })
+                .Select(g => new { Key = g.Key, SumId = g.Sum(x => (long)x.Id) })
                 .ToList();
 
             // Act
             var result = (from x in _users
                           group x by x.IsAdministrator into g
-                          select new { Key = g.Key, SumId = g.Sum(x => x.Id) })
+                          select new { Key = g.Key, SumId = g.Sum(x => (long)x.Id) })
                          .ToList();
 
             // Assert
